+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+BUG FIXES:
+
+- Fixes [a bug](https://github.com/launchdarkly/terraform-provider-launchdarkly/issues/60) where attempts to create `resource_launchdarkly_feature_flag` variations with an empty string value were throwing a panic.
+
 NOTES:
 
 - The `launchdarkly_feature_flag_environment` resource and data source's `flag_fallthrough` argument has been deprecated in favor of `fallthrough`. Please update your config to use `fallthrough` in order to maintain compatibility with future versions.